+++ src/sdk/TacSdk.ts
@@ -1,5 +1,5 @@
 import { Address, address, beginCell, Cell, toNano } from '@ton/ton';
-import { ethers, keccak256, toUtf8Bytes } from 'ethers';
+import { ethers, keccak256, toUtf8Bytes, isAddress as isEthereumAddress } from 'ethers';
 import type { SenderAbstraction } from '../sender';
 // import structs
 import {
@@ -257,23 +257,28 @@ export class TacSdk {
         return AssetOpType.JettonBurn;
     }
 
-    private aggregateJettons(assets?: AssetBridgingData[]): {
+    private async aggregateJettons(assets?: AssetBridgingData[]): Promise<{
         jettons: JettonBridgingData[];
         crossChainTonAmount: number;
-    } {
+    }> {
         const uniqueAssetsMap: Map<string, number> = new Map();
         let crossChainTonAmount = 0;
 
-        assets?.forEach((asset) => {
-            if (asset.amount <= 0) return;
+        for await (const asset of assets ?? []) {
+            if (asset.amount <= 0) continue;
 
             if (asset.address) {
-                validateTVMAddress(asset.address);
-                uniqueAssetsMap.set(asset.address, (uniqueAssetsMap.get(asset.address) || 0) + asset.amount);
+                const jettonAddress = isEthereumAddress(asset.address)
+                    ? await this.getTVMTokenAddress(asset.address)
+                    : asset.address;
+
+                validateTVMAddress(jettonAddress);
+
+                uniqueAssetsMap.set(jettonAddress, (uniqueAssetsMap.get(jettonAddress) || 0) + asset.amount);
             } else {
                 crossChainTonAmount += asset.amount;
             }
-        });
+        };
 
         const jettons: JettonBridgingData[] = Array.from(uniqueAssetsMap.entries()).map(([address, amount]) => ({
             address,
@@ -359,7 +364,7 @@ export class TacSdk {
         sender: SenderAbstraction,
         assets?: AssetBridgingData[],
     ): Promise<TransactionLinker> {
-        const aggregatedData = this.aggregateJettons(assets);
+        const aggregatedData = await this.aggregateJettons(assets);
         const transactionLinkerShardCount = aggregatedData.jettons.length == 0 ? 1 : aggregatedData.jettons.length;
 
         const caller = sender.getSenderAddress();