+++ tests/bridgeData.ts
@@ -28,6 +28,7 @@ const bridgeDataRawSender = async () => {
     // create sender abstraction
     const mnemonic = process.env.TVM_MNEMONICS || ''; // 24 words mnemonic
     const sender = await SenderFactory.getSender({
+        network: Network.Testnet,
         version: WALLET_VERSION,
         mnemonic,
       });