+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unrealeased]
+
+### Changed
+
+- `address` field in `AssetBridgingData` can be either EVM or TVM address
+
 ## [0.3.6] - 2025-01-15
 
 ### Changed