+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.7] - 2025-01-20
+
+### Changed
+
+- `address` field in `AssetBridgingData` can be either EVM or TVM address
+- Method `SenderFactory.getSender` requires additional parameter `network` when creating wallet wrapper using mnemonic
+
 ## [0.3.6] - 2025-01-15
 
 ### Changed